+++ packages/cw-orch-mock/Cargo.toml
@@ -10,7 +10,6 @@ description = "Adapter for interacting with cw-multi-test via the cw-orchestrato
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [features]
-snapshot-testing = ["dep:insta"]
 
 [dependencies]
 cw-orch-core = { workspace = true }
@@ -18,13 +17,11 @@ cosmwasm-std = { workspace = true }
 cw-multi-test = { workspace = true }
 cw-utils = { workspace = true }
 serde = { workspace = true }
-insta = { version = "1.34.0", features = ["yaml"], optional = true }
 
 [dev-dependencies]
 speculoos = { workspace = true }
 cw20 = { workspace = true }
 cw20-base = { workspace = true }
-cw-orch-mock = { workspace = true, features = ["snapshot-testing"] }
 
 mock-contract = { path = "../../contracts/mock_contract", features = [
   "interface",
